refactor: update page imports to use new @/ component paths
Updated imports in page files to use the new @/ path aliases:

Browse pages:
- molecules/page.tsx: MoleculeDisplay, ErrorState, BrowseTabs
- facilities/page.tsx: FacilityCardCompact, ErrorState, BrowseTabs

Detail pages:
- molecules/[id]/page.tsx: MoleculeDisplay, LoadingState, ErrorState
- facilities/[id]/page.tsx: ErrorState, LoadingState
- users/[id]/page.tsx: LoadingState, ErrorState, MoleculeDisplay, icons

Contribute pages:
- contribute/page.tsx: SignInButton
- contribute/molecule/page.tsx: Button, MoleculeDisplay, FieldTooltip, icons
- contribute/facility/page.tsx: Button

Auth pages:
- sign-in/page.tsx: SocialSignInButtons
- @modal/sign-in/page.tsx: SocialSignInButtons

TypeScript and production build pass successfully.
